+++ deepchem/models/__init__.py
@@ -47,7 +47,7 @@ try:
   from deepchem.models.torch_models import Pagtn, PagtnModel
   from deepchem.models.fcnet import MultitaskRegressor, MultitaskClassifier, MultitaskFitTransformRegressor
   from deepchem.models.torch_models import MEGNetModel
-  from deepchem.models.torch_models import CNN, CNNModule
+  from deepchem.models.torch_models import CNNModule
 except ModuleNotFoundError as e:
   logger.warning(
       f'Skipped loading some PyTorch models, missing a dependency. {e}')
