+++ deepchem/models/torch_models/__init__.py
@@ -9,4 +9,5 @@ from deepchem.models.torch_models.lcnn import LCNN, LCNNModel
 from deepchem.models.torch_models.pagtn import Pagtn, PagtnModel
 from deepchem.models.torch_models.mat import MAT, MATModel
 from deepchem.models.torch_models.megnet import MEGNetModel
-from deepchem.models.torch_models.cnn import CNN, CNNModule
+from deepchem.models.torch_models.layers import CNNModule
+from deepchem.models.torch_models.cnn import CNN
