+++ docs/source/api_reference/layers.rst
@@ -122,6 +122,9 @@ Keras Layers
 Torch Layers
 ------------
 
+.. autoclass:: deepchem.models.torch_models.layers.CNNModule
+  :members:
+
 .. autoclass:: deepchem.models.torch_models.layers.ScaleNorm
   :members:
