+++ docs/source/api_reference/models.rst
@@ -266,18 +266,6 @@ WGAN
 .. autoclass:: deepchem.models.WGAN
   :members:
 
-CNNModule
----
-
-.. autoclass:: deepchem.models.CNNModule
-  :members:
-
-CNN
----
-
-.. autoclass:: deepchem.models.CNN
-  :members:
-
 TextCNNModel
 ------------
 
@@ -331,6 +319,12 @@ TorchModel
 You can wrap an arbitrary :code:`torch.nn.Module` in a :code:`TorchModel` object.
 
 .. autoclass:: deepchem.models.TorchModel
+  :members:
+
+CNN
+---
+
+.. autoclass:: deepchem.models.CNN
   :members:
 
 MultitaskRegressor
